Update configure_katalon_studio_web_automation_test_project.md
English to Russian translation of the "Configuring Katalon Studio for Web Testing" page.
The Chrome addon Katalon utility is no more available in the Chrome webstore. So i have used in translation the Katalon Recorder addon and URL for it. The screenshot must be updated too. I think in English version it needs to be updated as well.